refactor: Extract YouTube comment utilities and action button definitions into dedicated files.
