+++ src/components/buttons/ActionButtonsFP.svelte
@@ -9,40 +9,9 @@
 
   let { onActionClick, isYouTubeActive } = $props()
 
-  const actions = [
-    {
-      key: 'analyze',
-      label: 'Analyze',
-      icon: 'heroicons:chart-bar-16-solid',
-      description: 'Analyze content structure and arguments',
-    },
-    {
-      key: 'explain',
-      label: 'Explain',
-      icon: 'heroicons:light-bulb-16-solid',
-      description: 'Explain in simple terms',
-    },
-    {
-      key: 'debate',
-      label: 'Debate',
-      icon: 'heroicons:scale-16-solid',
-      description: 'Analyze from multiple perspectives',
-    },
-    {
-      key: 'chapters',
-      label: 'Chapters',
-      icon: 'heroicons:queue-list-16-solid',
-      description: 'Summarize by chapters',
-      showOnlyForYouTube: true,
-    },
-    {
-      key: 'comments',
-      label: 'Comments',
-      icon: 'heroicons:chat-bubble-bottom-center-text-16-solid',
-      description: 'Analyze YouTube comments',
-      showOnlyForYouTube: true,
-    },
-  ]
+  import { ACTION_BUTTONS } from '@/lib/constants/actionConstants.js'
+
+  const actions = ACTION_BUTTONS
 
   // Filter actions based on current page type
   let visibleActions = $derived(
