+++ src/components/buttons/ActionButtonsMini.svelte
@@ -11,40 +11,9 @@
   import { Tooltip as BitsTooltip } from 'bits-ui'
   import { t } from 'svelte-i18n'
 
-  const actions = [
-    {
-      key: 'analyze',
-      label: 'Analyze',
-      icon: 'heroicons:chart-bar-16-solid',
-      description: 'Analyze content structure and arguments',
-    },
-    {
-      key: 'explain',
-      label: 'Explain',
-      icon: 'heroicons:light-bulb-16-solid',
-      description: 'Explain in simple terms',
-    },
-    {
-      key: 'debate',
-      label: 'Debate Mode',
-      icon: 'heroicons:scale-16-solid',
-      description: 'Analyze from multiple perspectives',
-    },
-    {
-      key: 'chapters',
-      label: 'Chapters',
-      icon: 'heroicons:queue-list-16-solid',
-      description: 'Summarize by chapters',
-      showOnlyForYouTube: true,
-    },
-    {
-      key: 'comments',
-      label: 'Comments',
-      icon: 'heroicons:chat-bubble-bottom-center-text-16-solid',
-      description: 'Summarize YouTube comments',
-      showOnlyForYouTube: true,
-    },
-  ]
+  import { ACTION_BUTTONS } from '@/lib/constants/actionConstants.js'
+
+  const actions = ACTION_BUTTONS
 
   async function handleActionClick(actionType) {
     if (actionType === 'chapters') {
