+++ src/lib/constants/actionConstants.js
@@ -1,0 +1,34 @@
+export const ACTION_BUTTONS = [
+  {
+    key: 'analyze',
+    label: 'Analyze',
+    icon: 'heroicons:chart-bar-16-solid',
+    description: 'Analyze content structure and arguments',
+  },
+  {
+    key: 'explain',
+    label: 'Explain',
+    icon: 'heroicons:light-bulb-16-solid',
+    description: 'Explain in simple terms',
+  },
+  {
+    key: 'debate',
+    label: 'Debate',
+    icon: 'heroicons:scale-16-solid',
+    description: 'Analyze from multiple perspectives',
+  },
+  {
+    key: 'chapters',
+    label: 'Chapters',
+    icon: 'heroicons:queue-list-16-solid',
+    description: 'Summarize by chapters',
+    showOnlyForYouTube: true,
+  },
+  {
+    key: 'comments',
+    label: 'Comments',
+    icon: 'heroicons:chat-bubble-bottom-center-text-16-solid',
+    description: 'Summarize YouTube comments',
+    showOnlyForYouTube: true,
+  },
+]
