+++ src/stores/summaryStore.svelte.js
@@ -27,6 +27,10 @@ import {
 // Import Deep Dive store
 import { resetDeepDive } from './deepDiveStore.svelte.js'
 import { showBlockingModeToast } from '@/lib/utils/toastUtils.js'
+import {
+  formatCommentsForAI,
+  fetchYouTubeComments,
+} from '@/lib/utils/youtubeUtils.js'
 
 // --- State ---
 export const summaryState = $state({
@@ -1049,10 +1053,7 @@ export async function logSingleSummaryToHistory(content, title, url, typeLabel) 
   }
 }
 
-export async function logAllGeneratedSummariesToHistory() {
-  // Deprecated: This function is no longer used for new logic but kept for reference or potential cleanup
-  // We are now using logSingleSummaryToHistory for granular history tracking
-}
+
 
 
 /**
@@ -1198,79 +1199,6 @@ export async function executeCustomAction(actionType) {
   }
 }
 
-/**
- * Format comments data for AI analysis
- * Cleans and structures comment data to optimize for AI processing
- * @param {Array} comments - Raw comment array from content script
- * @param {Object} metadata - Metadata object with video info
- * @returns {string} Formatted markdown string ready for AI
- */
-function formatCommentsForAI(comments, metadata) {
-  if (!comments || comments.length === 0) {
-    return 'No comments available.'
-  }
-
-  // Ultra-compact format - no metadata header
-  let formatted = ''
-  const seenTexts = new Set()
-
-  for (const comment of comments) {
-    // Skip duplicates and spam
-    const normalizedText = comment.text.trim().toLowerCase()
-    if (seenTexts.has(normalizedText) || comment.text.trim().length < 3) {
-      continue
-    }
-    seenTexts.add(normalizedText)
-
-    // Clean text: remove excessive emojis
-    let cleanText = comment.text
-      .trim()
-      .replace(/[\u{1F300}-\u{1F9FF}]{6,}/gu, '[emoji]')
-
-    // Truncate long comments (max 400 chars for better token efficiency)
-    if (cleanText.length > 400) {
-      cleanText = cleanText.substring(0, 397) + '...'
-    }
-
-    // Ultra compact format: @author|time|likes↑|replies💬|text
-    formatted += `@${comment.author.name}`
-    if (comment.author.isChannelOwner) {
-      formatted += '👤'
-    }
-    formatted += `|${comment.publishedTime}|${comment.likeCount}↑`
-    if (comment.replyCount > 0) {
-      formatted += `|${comment.replyCount}💬`
-    }
-    formatted += `|${cleanText}\n`
-
-    // Compact replies format
-    if (comment.replies && comment.replies.length > 0) {
-      for (const reply of comment.replies) {
-        // Clean reply text
-        let cleanReply = reply.text
-          .trim()
-          .replace(/[\u{1F300}-\u{1F9FF}]{6,}/gu, '[emoji]')
-
-        // Truncate long replies (max 200 chars)
-        if (cleanReply.length > 200) {
-          cleanReply = cleanReply.substring(0, 197) + '...'
-        }
-
-        // Skip very short replies
-        if (cleanReply.length < 3) {
-          continue
-        }
-
-        // Compact reply: indent + number.@author|text
-        formatted += `  ${reply.index}.@${reply.author.name}|${cleanReply}\n`
-      }
-    }
-
-    formatted += '\n'
-  }
-
-  return formatted
-}
 
 /**
  * Fetch and analyze YouTube comments
@@ -1349,10 +1277,9 @@ export async function fetchCommentSummary(
     console.log('[fetchCommentSummary] Sending message to content script...')
 
     // Send message to content script to fetch comments
-    const response = await browser.tabs.sendMessage(tabInfo.id, {
-      action: 'fetchYouTubeComments',
-      maxComments: maxComments,
-      maxRepliesPerComment: maxRepliesPerComment,
+    const response = await fetchYouTubeComments(tabInfo.id, {
+      maxComments,
+      maxRepliesPerComment,
     })
 
     console.log('[fetchCommentSummary] Received response:', response)
